+++ Cylinder.h
@@ -5,20 +5,20 @@
 #ifndef MYHOMEWORK_CYLINDER_H
 #define MYHOMEWORK_CYLINDER_H
 #include "Shape3D.h"
-#include "fstream"
+//#include "fstream"
 #include "string"
 #include "vector"
 
 class Cylinder : public Shape3D {
 private:
-    double radius;
-    double height;
+    float radius;
+    float height;
     int facets;
     vector<Shape3D::vertex> tops;
     vector<Shape3D::vertex> bottoms;
     //ofstream f;
 public:
-    Cylinder(double x, double y, double z, double r, double h, int f);
+    Cylinder(double x, double y, double z, float r, float h, int f);
     ~Cylinder();
     void print(string filename) override;
     void createVertices(vector<Shape3D::vertex>& vertices) override;
